refactor(ai-detection): split ScanDetailsPage into co-located modules
ScanDetailsPage had grown to 4225 lines, mixing row rendering, tab
panels, config tables, provider icons and pagination state in one file,
which made any change to a single tab hard to review in isolation.

Extract 15 focused modules under ScanDetails/ and leave the page with
scan/param handling, tab wiring and composition only, taking it from
4225 to 889 lines (-79%).

Also add the DE/FR/ES entries for the Secrets tab warning. The copy is
unchanged, but moving it from a multi-line JSX node into a single-line
prop made it newly visible to the i18n extractor, which would otherwise
fail the strict audit in CI.

No behaviour or UX changes.
