+++ Clients/src/presentation/pages/AIDetection/ScanDetails/VulnerabilityFindingRow.tsx
@@ -1,0 +1,380 @@
+/**
+ * @fileoverview Row component for LLM-specific vulnerability findings.
+ *
+ * @module pages/AIDetection/ScanDetails/VulnerabilityFindingRow
+ */
+
+import { useState } from "react";
+import { Box, Collapse, IconButton, Tooltip, Typography } from "@mui/material";
+import { ChevronDown, ChevronRight, EyeOff, FileCode, Link2, MoreHorizontal } from "lucide-react";
+import Chip from "../../../components/Chip";
+import { updateFindingGovernanceStatus } from "../../../../application/repository/aiDetection.repository";
+import SuppressFindingDialog from "../components/SuppressFindingDialog";
+import { Finding, GovernanceStatus } from "../../../../domain/ai-detection/types";
+import { palette } from "../../../themes/palette";
+import { FilePathItem } from "./FilePathItem";
+import { FindingRowGovernancePopover } from "./FindingRowGovernancePopover";
+import {
+  GOVERNANCE_STATUS_CONFIG,
+  isFindingSuppressed,
+  RISK_LEVEL_CONFIG,
+  VULN_TYPE_LABELS,
+} from "./scanDetailsConfig";
+
+// ============================================================================
+// Vulnerability Finding Row Component
+// ============================================================================
+
+interface VulnerabilityFindingRowProps {
+  finding: Finding;
+  repositoryOwner: string;
+  repositoryName: string;
+  scanId: number;
+  onGovernanceChange?: (findingId: number, status: GovernanceStatus | null) => void;
+  onStatusMessage?: (variant: "success" | "error", body: string) => void;
+}
+
+export function VulnerabilityFindingRow({
+  finding,
+  repositoryOwner,
+  repositoryName,
+  scanId,
+  onGovernanceChange,
+  onStatusMessage,
+}: VulnerabilityFindingRowProps) {
+  const [expanded, setExpanded] = useState(false);
+  const [governanceAnchor, setGovernanceAnchor] = useState<HTMLElement | null>(null);
+  const [localStatus, setLocalStatus] = useState<GovernanceStatus | null>(
+    finding.governance_status || null,
+  );
+  const [isUpdating, setIsUpdating] = useState(false);
+  const [suppressDialogOpen, setSuppressDialogOpen] = useState(false);
+  const vulnMeta = VULN_TYPE_LABELS[finding.finding_type] || {
+    label: finding.finding_type,
+    owaspId: "",
+  };
+
+  const getFileUrl = (filePath: string, lineNumber: number | null): string | null => {
+    if (!repositoryOwner || !repositoryName) return null;
+    const baseUrl = `https://github.com/${repositoryOwner}/${repositoryName}/blob/HEAD/${filePath}`;
+    return lineNumber ? `${baseUrl}#L${lineNumber}` : baseUrl;
+  };
+
+  const handleGovernanceClick = (event: React.MouseEvent<HTMLElement>) => {
+    event.stopPropagation();
+    setGovernanceAnchor(event.currentTarget);
+  };
+
+  const handleGovernanceClose = () => {
+    setGovernanceAnchor(null);
+  };
+
+  const handleStatusChange = async (newStatus: GovernanceStatus | null) => {
+    handleGovernanceClose();
+    if (newStatus === localStatus) return;
+
+    setIsUpdating(true);
+    try {
+      await updateFindingGovernanceStatus(scanId, finding.id, newStatus);
+      setLocalStatus(newStatus);
+      onGovernanceChange?.(finding.id, newStatus);
+      const statusLabel = newStatus ? GOVERNANCE_STATUS_CONFIG[newStatus].label : "unreviewed";
+      onStatusMessage?.("success", `Status updated to ${statusLabel}`);
+    } catch {
+      setLocalStatus(finding.governance_status || null);
+      onStatusMessage?.("error", "Failed to update governance status");
+    } finally {
+      setIsUpdating(false);
+    }
+  };
+
+  const StatusIcon = localStatus ? GOVERNANCE_STATUS_CONFIG[localStatus].icon : MoreHorizontal;
+  const statusColor = localStatus
+    ? GOVERNANCE_STATUS_CONFIG[localStatus].color
+    : palette.text.tertiary;
+
+  return (
+    <Box
+      sx={{
+        border: `1px solid ${palette.border.light}`,
+        borderRadius: "4px",
+        mb: "8px",
+        backgroundColor: palette.background.main,
+        opacity: isFindingSuppressed(finding) ? 0.6 : 1,
+      }}
+    >
+      {/* Header */}
+      <Box
+        sx={{
+          "display": "flex",
+          "alignItems": "center",
+          "p": "8px",
+          "cursor": "pointer",
+          "&:hover": {
+            backgroundColor: palette.background.accent,
+          },
+        }}
+        onClick={() => setExpanded(!expanded)}
+      >
+        <IconButton size="small" sx={{ mr: 1 }}>
+          {expanded ? <ChevronDown size={16} /> : <ChevronRight size={16} />}
+        </IconButton>
+
+        <Box sx={{ display: "flex", alignItems: "center", flex: 1 }}>
+          <Box>
+            <Typography sx={{ fontSize: "13px", fontWeight: 500 }}>
+              {finding.description || finding.name}
+            </Typography>
+            {finding.data_flow_summary && (
+              <Typography sx={{ fontSize: "13px", color: palette.text.tertiary, mt: 0.5 }}>
+                {finding.data_flow_summary}
+              </Typography>
+            )}
+          </Box>
+        </Box>
+
+        <Box sx={{ display: "flex", alignItems: "center", gap: "8px" }}>
+          {/* Suppressed Badge */}
+          {isFindingSuppressed(finding) && (
+            <Tooltip
+              title={
+                finding.suppressed ? "Matched a suppression rule" : "Manually marked as suppressed"
+              }
+              arrow
+              placement="top"
+            >
+              <Box
+                sx={{
+                  px: "8px",
+                  py: "2px",
+                  borderRadius: "4px",
+                  backgroundColor: palette.background.accent,
+                  border: `1px solid ${palette.border.light}`,
+                  display: "flex",
+                  alignItems: "center",
+                  gap: "4px",
+                }}
+              >
+                <EyeOff size={12} color={palette.text.tertiary} />
+                <Typography
+                  sx={{ fontSize: "12px", fontWeight: 500, color: palette.text.tertiary }}
+                >
+                  Suppressed
+                </Typography>
+              </Box>
+            </Tooltip>
+          )}
+          {/* Risk Level Badge */}
+          {finding.risk_level && (
+            <Tooltip
+              title={RISK_LEVEL_CONFIG[finding.risk_level]?.tooltip || ""}
+              arrow
+              placement="top"
+            >
+              <Box
+                sx={{
+                  px: "8px",
+                  py: "2px",
+                  borderRadius: "4px",
+                  backgroundColor: RISK_LEVEL_CONFIG[finding.risk_level]?.bgColor,
+                  border: `1px solid ${RISK_LEVEL_CONFIG[finding.risk_level]?.color}20`,
+                }}
+              >
+                <Typography
+                  sx={{
+                    fontSize: "12px",
+                    fontWeight: 500,
+                    color: RISK_LEVEL_CONFIG[finding.risk_level]?.color,
+                  }}
+                >
+                  {RISK_LEVEL_CONFIG[finding.risk_level]?.label}
+                </Typography>
+              </Box>
+            </Tooltip>
+          )}
+          {/* OWASP Badge */}
+          {vulnMeta.owaspId && (
+            <Box
+              sx={{
+                px: "8px",
+                py: "2px",
+                borderRadius: "4px",
+                backgroundColor: palette.accent.indigo.bg,
+                border: `1px solid ${palette.accent.indigo.text}20`,
+                display: "flex",
+                alignItems: "center",
+                gap: "4px",
+              }}
+            >
+              <Typography
+                sx={{
+                  fontSize: "12px",
+                  fontWeight: 500,
+                  color: palette.accent.indigo.text,
+                }}
+              >
+                {vulnMeta.owaspId}
+              </Typography>
+            </Box>
+          )}
+          {/* Cross-reference badge */}
+          {!!(
+            finding.vulnerability_details?.related_finding_types &&
+            (finding.vulnerability_details.related_finding_types as string[]).length > 0
+          ) && (
+            <Tooltip
+              title={`Also detected in: ${(
+                finding.vulnerability_details.related_finding_types as string[]
+              )
+                .map((t: string) => {
+                  const labels: Record<string, string> = {
+                    library: "Libraries",
+                    agent: "Agents",
+                    model_ref: "Models",
+                    api_call: "API calls",
+                    secret: "Secrets",
+                    rag_component: "RAG",
+                    dependency: "Dependencies",
+                  };
+                  return labels[t] || t;
+                })
+                .join(", ")} tab`}
+              arrow
+              placement="top"
+            >
+              <Box
+                sx={{
+                  px: "6px",
+                  py: "2px",
+                  borderRadius: "4px",
+                  backgroundColor: palette.accent.teal.bg,
+                  border: `1px solid ${palette.accent.teal.text}20`,
+                  display: "flex",
+                  alignItems: "center",
+                  gap: "4px",
+                }}
+              >
+                <Link2 size={10} color={palette.accent.teal.text} />
+                <Typography sx={{ fontSize: "11px", color: palette.accent.teal.text }}>
+                  Cross-ref
+                </Typography>
+              </Box>
+            </Tooltip>
+          )}
+          <Box sx={{ minWidth: 120, display: "flex", justifyContent: "center" }}>
+            <Chip label={vulnMeta.label} variant="default" size="small" />
+          </Box>
+          <Box
+            sx={{
+              display: "flex",
+              alignItems: "center",
+              gap: 0.5,
+              minWidth: 85,
+              justifyContent: "flex-end",
+            }}
+          >
+            <FileCode size={14} color={palette.text.tertiary} />
+            <Typography sx={{ fontSize: "13px", color: palette.text.tertiary }}>
+              {finding.file_count || finding.file_paths?.length || 0}{" "}
+              {(finding.file_count || finding.file_paths?.length || 0) === 1 ? "file" : "files"}
+            </Typography>
+          </Box>
+          {/* Governance Status Button */}
+          <Tooltip
+            title={
+              localStatus ? `Status: ${GOVERNANCE_STATUS_CONFIG[localStatus].label}` : "Set status"
+            }
+            arrow
+            placement="top"
+          >
+            <IconButton
+              size="small"
+              onClick={handleGovernanceClick}
+              disabled={isUpdating}
+              sx={{
+                "border": `1px solid ${palette.border.light}`,
+                "borderRadius": "4px",
+                "p": "4px",
+                "&:hover": { backgroundColor: palette.background.hover },
+              }}
+            >
+              <StatusIcon size={16} color={statusColor} />
+            </IconButton>
+          </Tooltip>
+        </Box>
+      </Box>
+
+      {/* Governance Status Popover */}
+      <FindingRowGovernancePopover
+        anchorEl={governanceAnchor}
+        localStatus={localStatus}
+        onClose={handleGovernanceClose}
+        onStatusChange={handleStatusChange}
+        onCreateSuppressionRule={() => {
+          handleGovernanceClose();
+          setSuppressDialogOpen(true);
+        }}
+      />
+
+      <SuppressFindingDialog
+        isOpen={suppressDialogOpen}
+        finding={finding}
+        onClose={() => setSuppressDialogOpen(false)}
+        onSuccess={(msg) => onStatusMessage?.("success", msg)}
+        onError={(msg) => onStatusMessage?.("error", msg)}
+      />
+
+      {/* Expanded Content */}
+      <Collapse in={expanded}>
+        <Box sx={{ p: "8px", borderTop: `1px solid ${palette.border.light}` }}>
+          {/* Mitigation */}
+          {finding.mitigation && (
+            <Box sx={{ mb: "8px" }}>
+              <Typography sx={{ fontSize: "13px", fontWeight: 500, mb: "4px" }}>
+                Recommended mitigation
+              </Typography>
+              <Typography sx={{ fontSize: "13px", color: palette.text.secondary }}>
+                {finding.mitigation}
+              </Typography>
+            </Box>
+          )}
+
+          {/* File paths */}
+          <Typography sx={{ fontSize: "13px", fontWeight: 500, mb: 1 }}>Found in:</Typography>
+          <Box
+            sx={{
+              maxHeight: 200,
+              overflow: "auto",
+              backgroundColor: palette.background.accent,
+              borderRadius: "4px",
+              p: 1,
+            }}
+          >
+            {finding.file_paths?.slice(0, 20).map((fp, idx) => (
+              <FilePathItem
+                key={idx}
+                path={fp.path}
+                lineNumber={fp.line_number}
+                matchedText={fp.matched_text}
+                fileUrl={getFileUrl(fp.path, fp.line_number)}
+              />
+            ))}
+            {(finding.file_paths?.length || 0) > 20 && (
+              <Typography
+                sx={{
+                  fontSize: "13px",
+                  color: palette.text.tertiary,
+                  fontStyle: "italic",
+                  mt: 1,
+                  px: 1,
+                }}
+              >
+                And {(finding.file_paths?.length || 0) - 20} more files...
+              </Typography>
+            )}
+          </Box>
+        </Box>
+      </Collapse>
+    </Box>
+  );
+}
